+++ app/upgrades/v12/upgrades.go
@@ -75,6 +75,8 @@ func CreateUpgradeHandler(
 		keepers.ICAHostKeeper.SetParams(ctx, hostParams)
 
 		// Initialize TWAP state
+		// N.B.: deprecation nolint
+		// nolint: staticcheck
 		latestPoolId := keepers.GAMMKeeper.GetNextPoolId(ctx) - 1
 		err := keepers.TwapKeeper.MigrateExistingPools(ctx, latestPoolId)
 		if err != nil {