+++ x/ibc-hooks/ibc_middleware_test.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"testing"
 
+	ibchooks "github.com/osmosis-labs/osmosis/v13/x/ibc-hooks"
 	wasmkeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	"github.com/osmosis-labs/osmosis/v13/x/gamm/pool-models/balancer"
 	gammtypes "github.com/osmosis-labs/osmosis/v13/x/gamm/types"
-	ibchooks "github.com/osmosis-labs/osmosis/v13/x/ibc-hooks"
 	minttypes "github.com/osmosis-labs/osmosis/v13/x/mint/types"
 
 	"github.com/osmosis-labs/osmosis/v13/osmoutils"